+++ db/migrate/20150610053320_add_on_time_out_to_asyncapi_client_job.rb
@@ -1,0 +1,5 @@
+class AddOnTimeOutToAsyncapiClientJob < ActiveRecord::Migration
+  def change
+    add_column :asyncapi_client_jobs, :on_time_out, :string
+  end
+end