+++ spec/dummy/db/migrate/20150611040554_add_on_time_out_to_asyncapi_client_job.asyncapi_client.rb
@@ -1,0 +1,6 @@
+# This migration comes from asyncapi_client (originally 20150610053320)
+class AddOnTimeOutToAsyncapiClientJob < ActiveRecord::Migration
+  def change
+    add_column :asyncapi_client_jobs, :on_time_out, :string
+  end
+end