+++ spec/fixtures/on_time_out.rb
@@ -1,0 +1,3 @@
+class OnTimeOut
+
+end