+++ spec/workers/job_status_worker_spec.rb
@@ -3,18 +3,20 @@ require "spec_helper"
 module Asyncapi::Client
   describe JobStatusWorker, "#perform" do
 
-    it "executes the appropriate callback class" do
-      job = build_stubbed(
-        :asyncapi_client_job,
-        on_queue: "OnQueue",
-        on_success: "OnSuccess",
-        status: :success,
-      )
+    described_class::STATUS_CALLBACK_MAP.each do |status, attr_name|
+      it "executes #{attr_name} callback class" do
+        callback_class_name = attr_name.to_s.classify
+        job = build_stubbed(
+          :asyncapi_client_job,
+          attr_name => callback_class_name,
+          status: status,
+        )
 
-      allow(Job).to receive(:find).with(job.id).and_return(job)
-      expect(OnSuccess).to receive(:call).with(job.id)
+        allow(Job).to receive(:find).with(job.id).and_return(job)
+        expect(callback_class_name.constantize).to receive(:call).with(job.id)
 
-      described_class.new.perform(job.id)
+        described_class.new.perform(job.id)
+      end
     end
 
     context "status callback is not defined" do